updated expression parsing; added range function
L14 is now left associative and contains all assignment based operations
except assignment (=) itself; that is contained in L15 and highest level
is L16 now; also updated many constructs to use L15 expression instead
of L13 or L14;
also, range function for integers has been added in library/utils.cpp and
tested on `tests/facto_iterate.fer` (utils is internal module, loaded
with core)

+++ src/VM/Vars/AttrBased.cpp
@@ -9,6 +9,6 @@
 
 #include "Base.hpp"
 
-var_attr_based_t::var_attr_based_t( const size_t & type, const size_t & src_id, const size_t & idx )
+var_attr_based_t::var_attr_based_t( const int & type, const size_t & src_id, const size_t & idx )
 	: var_base_t( type, src_id, idx ) {}
 var_attr_based_t::~var_attr_based_t() {}
